bid/requirements.txt
alembic==1.3.1
bcrypt==3.1.7
blinker==1.4
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==6.2.1
pycparser==2.19
python-dateutil==2.8.1
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.11
Werkzeug==0.16.0
WTForm==1.0
WTForms==2.2.1
